// src/gbp/nucleo/interfaces/product-storage-group-item.interface.ts
export interface ProductStorageGroupItem 
{  
  item_id: string;
  item_code: string;
  item_desc: string;
  item_detail: string;
  item_vendorCode: string;
  cat_id: string;
  subcat_id: string;
  brand_id: string;
  pres_id: string;
  item_upb: string;
  item_upp: string;
  item_cantMin: string;
  item_cantOpt: string;
  item_mult: string;
  item_commission: string;
  item_guarantee: string;
  item_expser: string;
  item_made: string;
  item_web: string;
  item_annotation: string;
  item_annotation1: string;
  item_annotation2: string;
  item_newness: string;
  item_liquidation: string;
  item_withDecimal: string;
  supp_id: string;
  item_qtyMin1: string;
  item_qtyMin2: string;
  item_qtyMin3: string;
  item_markup_1: string;
  item_markup_2: string;
  item_markup_3: string;
  item_markup_4: string;
  item_markup_5: string;
  item_markup_6: string;
  item_markup_7: string;
  item_markup_8: string;
  item_markup_9: string;
  item_markup_10: string;
  item_markup_11: string;
  item_markup_12: string;
  item_markup_13: string;
  item_markup_14: string;
  item_markup_15: string;
  item_markup_16: string;
  item_weight: string;
  item_not4Sale: string;
  item_disabled: string;
  item_disabledInBalance: string;
  mu_id: string;
  stock: string;
  tax_id: string;
  tax_percentage: string;
  tax_id_II: string;
  tax_percentage_II: string;
  imageExisting: string;
  subcataux_id: string;
  item_WebSite_desc: string;
  item_wide: string;
  item_large: string;
  item_higth: string;

  // Propiedades exclusiva del group item.
  PhisicalStock: string;
  option_id: string;
}